refactor(ops): extract serve tests to separate file
Move the test module from serve.rs (983 lines) into serve_tests.rs
(243 lines), bringing serve.rs down to 738 lines — well under the
800-line style guideline.
